histograms: initial traceeval alloc and dealloc
Initial implementation of struct traceeval and its constructor and
destructor functions, traceeval_init() and traceeval_release(). These
functions reply on other constructors and destructors (defined in this
commit) for the fields of struct traceeval.
Also defines the struct hist_table and it's counterpart struct entry.
These two structs constitute the histogram and the 'rows'/'entries' of
data that can be inserted.