histograms: use strncpy() instead of strlcpy()
strlcpy() is a safe replacement for strncpy() which requires the bsd
library. To reduce dependencies, the use of strlcpy() in type_alloc()
to copy the user defined type's name field is replaced with strncpy().
This reduces possible strain on low resource enviroments, but
requires users to null terminate their strings, which is common, good
practice. The docstrings have been updated to provide a warning where
it is relevant.