histograms: NULL vals, NONE names, realloc leak
traceeval_init() accpets NULL for the vals parameter. When NULL is
received for vals, the traceeval instance's def_vals field is set to a
single struct traceeval_type of type TRACEEVAL_TYPE_NONE. Since this
is the terminator type, the internal state treats the histogram as
if it cannot have values. This change is reflected in the
documentation.

All struct traceeval_type instances passed to
traceeval_init() are required to have a non-NULL name field except for
those of type TRACEEVAL_TYPE_NONE. For the instance of type
TRACEEVAL_TYPE_NONE, since this is the terminator of a list of types, it
need not be named. Therefore, any name associated with it is
disregarded.

In type_alloc(), if realloc() returned NULL, a memory leak would occur.
A temporary variable is used to validate the return of realloc(),
which solves the memory leak.