histograms: traceeval_dynamic distinguishing
To increase flexibility, a id field is added to struct traceeval_type.
The field is optional, and it's intended purpose is to distingush
between multiple struct traceeval_type instances of type
TRACEEVAL_TYPE_NONE. The signatures of dyn_cmp() and dyn_release() have
been updated to encorperate this flexibility by adding an argument for a
struct traceeval_type instance to let the user access the "definition"
of the dynamic data.